refactor: modularize service ambition generator
